requirements.txt
fastapi 
uvicorn
google-genai
dotenv
pillow
python-multipart
torch
transformers